Properties defined

+++ src/main/java/com/fitandfocused/model/WorkoutPlan.java
@@ -3,6 +3,8 @@ package com.fitandfocused.model;
 import lombok.Data;
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Data
 @Table(name = "tb_workout_plan")
@@ -14,5 +16,5 @@ public class WorkoutPlan {
     private String dayOfWeek;
     private String startTime;
     private String endTime;
-    private double pricePerHour;
+    private BigDecimal pricePerHour;
 }
